LONG!! implement import file to buffer
The main ideea of this work/patch was to implement fully the function
"buffer_from_file()". The file is read into chunks and then lines are
built that are inserted into the buffer.
Previously only the code for saving only one buffer was implemented.
That code portion was moved over to the function "buffer_from_string()".
Additionally some other functions/modifications were needed:
- buffer_append_line() now either appends the whole line at the end of
  the line list or attaches the data of the line to the last line if the
  last line in the buffer's line list was not properly terminated (with
  a '\n')
- line_append_string() takes a line and a string and appends the string
  to the line expanding the line if needed.
  CAUTION: the input line should not be used after calling this
  function. You should always use this function as:
          line = line_append_string(line, text);
- additional line functions have been introduced
- the line struct now contains 'prev'; this pointer help both with
  insertion and with navigation.

Signed-off-by: Radu Alexe <alexeradu2007@gmail.com>